Coal course project
In the brick breaker game, the player moves a paddle from side to side to hit a BALL. The game
objective is to eliminate all of the BRICKS at the top of the screen by hitting them with the BALL.
But if the ball hit the bottom ENCLOSURE, the player loses and the game ends! To win the game,
all the BRICKS must be eliminated. The game is split into many levels, which must be completed
in sequence. There will be a time limit of 4 minutes and the remaining time will be shown with the
counter. The purpose of this game is to complete all the levels without losing all lives. The player
will have a maximum of 3 lives